# docs/swagger.yaml
basePath: /api
definitions:
  dto.AddCommentRequest:
    properties:
      comment:
        properties:
          body:
            type: string
        required:
        - body
        type: object
    type: object
  dto.ArticleBody:
    properties:
      author:
        $ref: '#/definitions/dto.ProfileBody'
      body:
        type: string
      createdAt:
        type: string
      description:
        type: string
      favorited:
        type: boolean
      favoritesCount:
        type: integer
      slug:
        type: string
      tagList:
        items:
          type: string
        type: array
      title:
        type: string
      updatedAt:
        type: string
    type: object
  dto.ArticleResponse:
    properties:
      article:
        $ref: '#/definitions/dto.ArticleBody'
    type: object
  dto.ArticlesResponse:
    properties:
      articles:
        items:
          $ref: '#/definitions/dto.ArticleBody'
        type: array
      articlesCount:
        type: integer
    type: object
  dto.CommentBody:
    properties:
      author:
        $ref: '#/definitions/dto.ProfileBody'
      body:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      updatedAt:
        type: string
    type: object
  dto.CommentResponse:
    properties:
      comment:
        $ref: '#/definitions/dto.CommentBody'
    type: object
  dto.CommentsResponse:
    properties:
      comments:
        items:
          $ref: '#/definitions/dto.CommentBody'
        type: array
    type: object
  dto.CreateArticleRequest:
    properties:
      article:
        properties:
          body:
            type: string
          description:
            maxLength: 1000
            minLength: 1
            type: string
          tagList:
            items:
              type: string
            type: array
          title:
            maxLength: 500
            minLength: 1
            type: string
        required:
        - body
        - description
        - title
        type: object
    type: object
  dto.LoginRequest:
    properties:
      user:
        properties:
          email:
            type: string
          password:
            type: string
        required:
        - email
        - password
        type: object
    type: object
  dto.ProfileBody:
    properties:
      bio:
        type: string
      following:
        type: boolean
      image:
        type: string
      username:
        type: string
    type: object
  dto.ProfileResponse:
    properties:
      profile:
        $ref: '#/definitions/dto.ProfileBody'
    type: object
  dto.RegisterRequest:
    properties:
      user:
        properties:
          email:
            type: string
          password:
            minLength: 8
            type: string
          username:
            maxLength: 100
            minLength: 2
            type: string
        required:
        - email
        - password
        - username
        type: object
    type: object
  dto.TagsResponse:
    properties:
      tags:
        items:
          type: string
        type: array
    type: object
  dto.UpdateArticleRequest:
    properties:
      article:
        properties:
          body:
            type: string
          description:
            type: string
          title:
            type: string
        type: object
    type: object
  dto.UserBody:
    properties:
      bio:
        type: string
      email:
        type: string
      image:
        type: string
      token:
        type: string
      username:
        type: string
    type: object
  dto.UserResponse:
    properties:
      user:
        $ref: '#/definitions/dto.UserBody'
    type: object
host: localhost:8080
info:
  contact:
    name: API Support
  description: Conduit backend implementation using Go + Echo + GORM (PostgreSQL).
  title: RealWorld Echo API
  version: "1.0"
paths:
  /articles:
    get:
      parameters:
      - description: Filter by tag
        in: query
        name: tag
        type: string
      - description: Filter by author username
        in: query
        name: author
        type: string
      - description: Filter by username who favorited
        in: query
        name: favorited
        type: string
      - description: Limit (default 20)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ArticlesResponse'
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: List articles
      tags:
      - articles
    post:
      consumes:
      - application/json
      parameters:
      - description: Article data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateArticleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ArticleResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create article
      tags:
      - articles
  /articles/{slug}:
    delete:
      parameters:
      - description: Article slug
        in: path
        name: slug
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete article
      tags:
      - articles
    get:
      parameters:
      - description: Article slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ArticleResponse'
        "404":
          description: Article not found
          schema:
            additionalProperties: true
            type: object
      summary: Get article
      tags:
      - articles
    put:
      consumes:
      - application/json
      parameters:
      - description: Article slug
        in: path
        name: slug
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateArticleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ArticleResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update article
      tags:
      - articles
  /articles/{slug}/comments:
    get:
      parameters:
      - description: Article slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CommentsResponse'
        "404":
          description: Article not found
          schema:
            additionalProperties: true
            type: object
      summary: Get comments
      tags:
      - articles
    post:
      consumes:
      - application/json
      parameters:
      - description: Article slug
        in: path
        name: slug
        required: true
        type: string
      - description: Comment body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AddCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CommentResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Article not found
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add comment
      tags:
      - articles
  /articles/{slug}/comments/{id}:
    delete:
      parameters:
      - description: Article slug
        in: path
        name: slug
        required: true
        type: string
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Comment not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete comment
      tags:
      - articles
  /articles/{slug}/favorite:
    delete:
      parameters:
      - description: Article slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ArticleResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Article not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Unfavorite article
      tags:
      - articles
    post:
      parameters:
      - description: Article slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ArticleResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Article not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Favorite article
      tags:
      - articles
  /articles/feed:
    get:
      parameters:
      - description: Limit (default 20)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ArticlesResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get feed
      tags:
      - articles
  /profiles/{username}:
    get:
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProfileResponse'
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
      summary: Get profile
      tags:
      - profiles
  /profiles/{username}/follow:
    delete:
      parameters:
      - description: Username to unfollow
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProfileResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Unfollow user
      tags:
      - profiles
    post:
      parameters:
      - description: Username to follow
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProfileResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Follow user
      tags:
      - profiles
  /tags:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TagsResponse'
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: List all tags
      tags:
      - tags
  /user:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - users
  /users:
    post:
      consumes:
      - application/json
      parameters:
      - description: User credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "422":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - users
  /users/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "401":
          description: Invalid credentials
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
      summary: Login and receive a JWT token
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: 'Use format: Token <jwt>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"